+++ src/frontend/web_application/src/scenes/Contact/index.js
@@ -2,6 +2,7 @@ import { createSelector } from 'reselect';
 import { bindActionCreators, compose } from 'redux';
 import { connect } from 'react-redux';
 import { withTranslator } from '@gandi/react-translate';
+import { createNotification, NOTIFICATION_TYPE_ERROR } from 'react-redux-notify';
 import { requestContact, updateContact } from '../../store/modules/contact';
 import Presenter from './presenter';
 
@@ -20,6 +21,10 @@ const mapStateToProps = createSelector(
 const mapDispatchToProps = dispatch => bindActionCreators({
   requestContact,
   updateContact,
+  notifyError: message => createNotification({
+    message,
+    type: NOTIFICATION_TYPE_ERROR,
+  }),
 }, dispatch);
 
 export default compose(